BackEnd-app/requirements.txt
# Backend Framework
fastapi==0.95.2
uvicorn==0.23.2

# # Database Libraries
sqlalchemy==1.4.49
psycopg2-binary==2.9.8
pymongo        # MongoDB support

# AI Libraries
pymupdf
langchain
langchain_groq
langchain-text-splitters
langchain_community
langchain_pinecone
fastembed
nest-asyncio
unstructured[pdf]
python-dotenv
